+++ apps/web/src/app/auth/layout.tsx
@@ -1,4 +1,4 @@
-import TopNavigationBar from '@/components/layout/nav/top/TopNavigationBar';
+import TopNavigationBar from '@/components/layout/navbar/TopNavigationBar';
 
 interface AuthLayoutProps {
   children?: React.ReactNode;
@@ -8,7 +8,7 @@ export default function AuthLayout({ children }: AuthLayoutProps) {
   return (
     <div className="w-full min-h-screen flex flex-col">
       <div>
-        <TopNavigationBar />
+        <TopNavigationBar showSidebarTrigger={false} />
       </div>
 
       <div className="grow flex items-center justify-center">{children}</div>
