+++ apps/web/src/components/layout/navbar/_components/ProjectSwitcher.tsx
