+++ apps/web/src/components/layout/navbar/_components/SearchBar.tsx
